Published latest changes to the sample

+++ entities/Folder.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 
 @SuppressWarnings("serial")
 public class Folder implements Serializable {
-	public int SyncpointId;
+	public long SyncpointId;
 
-	public int FolderId;
+	public long FolderId;
 
 	public String VirtualPath;
 
@@ -20,9 +20,9 @@ public class Folder implements Serializable {
 
 	/** 
 	 * Returns a logging-friendly string
-	 * @return A <see cref="System.String"/>
 	 */
 	public final String ToLoggingString() {
-		return String.format("Virtual Path: %1$s, Name: %2$s, Status: %3$s, VirtualFolderId: %4$s, DataFolderId: %5$s", VirtualPath, Name, Status.toString(), SyncpointId, FolderId);
+		return String.format("Virtual Path: %1$s, Name: %2$s, Status: %3$s, VirtualFolderId: %4$s, DataFolderId: %5$s",
+				VirtualPath, Name, Status.toString(), SyncpointId, FolderId);
 	}
 }
